תיקון סוג משתנה

+++ FVCalc.java
@@ -3,8 +3,8 @@ public class FVCalc {
 	public static void main(String[] args){
 		// Replace this comment with your code
 		double currentValue = Double.parseDouble (args[0]);
-		int rate = Integer.parseInt (args[1]);
-		int n = Integer.parseInt (args[2]);
+		double rate = Double.parseDouble (args[1]);
+		double n = Double.parseDouble (args[2]);
 
 
 		double futureValueDouble = currentValue*(Math.pow(1+(rate/100.0),n)) ;
